requirements.txt
anthropic>=0.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
python-dotenv>=1.0.0
